Attachments and Image Selector wizards - upload-button does not upload pictures and files #1198
Set option that was added in the new version of the file uploader (`requireSuccessJson`) to `false`, since we don't follow specific JSON response scheme and don't add `success` parameter to that JSON, absence of which has lead to request treated as failed fy the library.
Added localization for attachment upload success and failure.